backend/requirements.txt
# fastapi==0.109.2
# uvicorn[standard]==0.27.1
# sqlalchemy==2.0.27
# psycopg2-binary==2.9.9
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# python-multipart==0.0.9
# pydantic==2.6.4
# pydantic-settings==2.2.1
# python-dotenv==1.0.1
# bcrypt==4.1.2
# email-validator==2.1.1
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
bcrypt==4.1.2
email-validator==2.1.1